Got movement working (finally :/)
Changed to using numpy arrays instead of tuples so that proper arithmetic can be done

Added ability for velocities to be drawn on the segments as a debugging tool

Velocities now use pixel values instead of being 0-1 directional vectors. This is to consolidate the number of parameters passed into functions.

KNOWN ISSUES
Snake only turns left at the moment even though the code should produce a random result
Segments do not update to the correct turning type, they stay as Straight segments